flask_app/requirements.txt
flask
gunicorn==19.5.0
psycopg2-binary
pandas